Simplify Gemfile
We shouldn't be using `master` branch as the gemspec specifies what
version of Solidus this is compat with. Also the mysql and pg gems
are not needed as SQLite is used for testing.

This this I can run the tests. There are two feature tests that still
fail due to what I believe is a authorization problem in the mocked
test environment.